pyproject.toml
[project]
name = "credit-loan"
version = "0.1.0"
description = ""
authors = [
    {name = "Cracker051",email = "stalker1guppi@ukr.net"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "ruff (>=0.11.0,<0.12.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"

[tool.ruff]
line-length = 120
lint.extend-ignore = ["E203"]
